+++ src/Field.js
@@ -6,40 +6,42 @@ import {
   $try,
   $hasFiles,
   $isBool,
-  $isEvent } from './utils';
+  $isEvent,
+  pathToStruct } from './utils';
 
 import {
   parseInput,
-  parseGetLabel,
+  parseCheckOutput,
+  defaultValue,
   defaultClearValue } from './parser';
 
 const setupFieldProps = (instance, props, data) =>
   Object.assign(instance, {
-    $value: instance.$initial,
-    $label: props.$label || data.label || instance.name,
-    $placeholder: props.$placeholder || data.placeholder || '',
-    $disabled: props.$disabled || data.disabled || false,
-    $bindings: props.$bindings || data.bindings || 'default',
-    $related: props.$related || data.related || [],
-    $validators: toJS(props.$validators || data.validators || null),
-    $validatedWith: props.$validatedWith || data.validatedWith || 'value',
-    $rules: props.$rules || data.rules || null,
-    $observers: props.$observers || data.observers || null,
-    $interceptors: props.$interceptors || data.interceptors || null,
-    $extra: props.$extra || data.extra || null,
-    $options: props.$options || data.options || {},
-    $hooks: props.$hooks || data.hooks || {},
-    $handlers: props.$handlers || data.handlers || {},
+    $label: props.$label || data && data.label || '',
+    $placeholder: props.$placeholder || data && data.placeholder || '',
+    $disabled: props.$disabled || data && data.disabled || false,
+    $bindings: props.$bindings || data && data.bindings || 'default',
+    $related: props.$related || data && data.related || [],
+    $validators: toJS(props.$validators || data && data.validators || null),
+    $validatedWith: props.$validatedWith || data && data.validatedWith || 'value',
+    $rules: props.$rules || data && data.rules || null,
+    $observers: props.$observers || data && data.observers || null,
+    $interceptors: props.$interceptors || data && data.interceptors || null,
+    $extra: props.$extra || data && data.extra || null,
+    $options: props.$options || data && data.options || {},
+    $hooks: props.$hooks || data && data.hooks || {},
+    $handlers: props.$handlers || data && data.handlers || {},
   });
 
 const setupDefaultProp = (instance, data, props, update, {
-  isEmptyArray, checkArray,
+  isEmptyArray,
 }) => parseInput(instance.$input, {
+  nullable: true,
   isEmptyArray,
   type: instance.type,
-  unified: update ? '' : checkArray(data.default),
-  separated: checkArray(props.$default),
-  initial: checkArray(instance.$initial),
+  unified: update ? defaultValue({type: instance.type}) : data && data.default,
+  separated: props.$default,
+  fallback: instance.$initial,
 });
 
 export default class Field extends Base {
@@ -52,7 +54,6 @@ export default class Field extends Base {
   key;
   name;
   path;
-  type;
   state;
 
   $observers;
@@ -66,6 +67,7 @@ export default class Field extends Base {
 
   @observable $options;
   @observable $value;
+  @observable $type;
   @observable $label;
   @observable $placeholder;
   @observable $default;
@@ -82,9 +84,9 @@ export default class Field extends Base {
   @observable $focused = false;
   @observable $touched = false;
   @observable $changed = false;
+  @observable $blurred = false;
+  @observable $deleted = false;
 
-  @observable $submitting = false;
-  @observable $validating = false;
   @observable $clearing = false;
   @observable $resetting = false;
 
@@ -108,7 +110,7 @@ export default class Field extends Base {
     this.state = state;
 
     this.setupField(key, path, data, props, update);
-    this.checkDVRValidationPlugin();
+    this.checkValidationPlugins();
     this.initNestedFields(data, update);
 
     this.incremental = (this.hasIncrementalKeys !== 0);
@@ -163,15 +165,19 @@ export default class Field extends Base {
   }
 
   @computed get initial() {
-    return this.getComputedProp('initial');
+    return this.$initial
+      ? toJS(this.$initial)
+      : this.getComputedProp('initial');
+  }
+
+  @computed get default() {
+    return this.$default
+      ? toJS(this.$default)
+      : this.getComputedProp('default');
   }
 
   set initial(val) {
     this.$initial = parseInput(this.$input, { separated: val });
-  }
-
-  @computed get default() {
-    return this.getComputedProp('default');
   }
 
   set default(val) {
@@ -182,16 +188,12 @@ export default class Field extends Base {
     return (this.submitting || this.clearing || this.resetting);
   }
 
-  @computed get submitting() {
-    return toJS(this.$submitting);
-  }
-
-  @computed get validating() {
-    return toJS(this.$validating);
+  @computed get type() {
+    return toJS(this.$type);
   }
 
   @computed get label() {
-    return parseGetLabel(this.$label);
+    return toJS(this.$label);
   }
 
   @computed get placeholder() {
@@ -227,7 +229,7 @@ export default class Field extends Base {
   }
 
   @computed get validatedValue() {
-    return this[this.$validatedWith];
+    return parseCheckOutput(this, this.$validatedWith)
   }
 
   @computed get error() {
@@ -245,22 +247,27 @@ export default class Field extends Base {
       && this.check('isValid', true);
   }
 
+  @computed get isDefault() {
+    return !_.isNil(this.default) &&
+      _.isEqual(this.default, this.value);
+  }
+
   @computed get isDirty() {
-    return this.hasNestedFields
-      ? this.check('isDirty', true)
-      : !_.isEqual(this.$default, this.value);
+    return !_.isUndefined(this.initial) &&
+      !_.isEqual(this.initial, this.value);
   }
 
   @computed get isPristine() {
-    return this.hasNestedFields
-      ? this.check('isPristine', true)
-      : _.isEqual(this.$default, this.value);
+    return !_.isNil(this.initial) &&
+      _.isEqual(this.initial, this.value) ;
   }
 
-  @computed get isDefault() {
-    return this.hasNestedFields
-      ? this.check('isDefault', true)
-      : _.isEqual(this.$default, this.value);
+  @computed get isEmpty() {
+    if (this.hasNestedFields) return this.check('isEmpty', true);
+    if (_.isBoolean(this.value)) return !!this.$value;
+    if (_.isNumber(this.value)) return false;
+    if (_.isDate(this.value)) return false;
+    return _.isEmpty(this.value);
   }
 
   @computed get resetting() {
@@ -275,18 +282,16 @@ export default class Field extends Base {
       : this.$clearing;
   }
 
-  @computed get isEmpty() {
-    if (this.hasNestedFields) return this.check('isEmpty', true);
-    if (_.isBoolean(this.value)) return !!this.$value;
-    if (_.isNumber(this.value)) return false;
-    if (_.isDate(this.value)) return false;
-    return _.isEmpty(this.value);
-  }
-
   @computed get focused() {
     return this.hasNestedFields
       ? this.check('focused', true)
       : this.$focused;
+  }
+
+  @computed get blurred() {
+    return this.hasNestedFields
+      ? this.check('blurred', true)
+      : this.$blurred;
   }
 
   @computed get touched() {
@@ -299,6 +304,12 @@ export default class Field extends Base {
     return this.hasNestedFields
       ? this.check('changed', true)
       : this.$changed;
+  }
+
+  @computed get deleted() {
+    return this.hasNestedFields
+      ? this.check('deleted', true)
+      : this.$deleted;
   }
 
   /* ------------------------------------------------------------------ */
@@ -339,6 +350,10 @@ export default class Field extends Base {
   onBlur = (...args) =>
     this.execHandler('onBlur', args,
       action(() => {
+        if (!this.$blurred) {
+          this.$blurred = true;
+        }
+
         this.$focused = false;
       }));
 
@@ -373,43 +388,44 @@ export const prototypes = {
     this.key = $key;
     this.path = $path;
     this.id = this.state.options.get('uniqueId').apply(this, [this]);
-    const isEmptyArray = (_.has($data, 'fields') && _.isArray($data.fields));
-    const checkArray = val => isEmptyArray ? [] : val;
+    const struct = this.state.struct();
+    const structPath = pathToStruct(this.path)
+    const isEmptyArray = Array.isArray(struct) ? 
+      struct.filter(s => s.startsWith(structPath)).find(s => s.substr(structPath.length, 2) === '[]')
+      : Array.isArray(_.get(struct, this.path))
 
-    const {
-      $value,
-      $type,
-      $input,
-      $output,
-    } = $props;
+    const { $type, $input, $output } = $props;
 
     // eslint-disable-next-line
-    if (_.isNil($data)) $data = '';
+    // if (_.isNil($data)) $data = '';
 
     if (_.isPlainObject($data)) {
-      const {
-        value,
-        type,
-        input,
-        output,
-      } = $data;
+      const { type, input, output } = $data;
 
-      this.type = $type || type || 'text';
       this.name = _.toString($data.name || $key);
-
+      this.$type = $type || type || 'text';
       this.$input = $try($input, input, this.$input);
       this.$output = $try($output, output, this.$output);
 
-      this.$initial = parseInput(this.$input, {
+      this.$value = parseInput(this.$input, {
         isEmptyArray,
         type: this.type,
-        unified: checkArray(value),
-        separated: checkArray($props.$initial),
-        initial: checkArray($data.initial),
+        unified: $data.value,
+        separated: $props.$value,
+        fallback: $props.$initial,
+      });
+
+      this.$initial = parseInput(this.$input, {
+        nullable: true,
+        isEmptyArray,
+        type: this.type,
+        unified: $data.initial,
+        separated: $props.$initial,
+        fallback: this.$value,
       });
 
       this.$default = setupDefaultProp(this, $data, $props, update, {
-        isEmptyArray, checkArray,
+        isEmptyArray,
       });
 
       setupFieldProps(this, $props, $data);
@@ -418,20 +434,28 @@ export const prototypes = {
 
     /* The field IS the value here */
     this.name = _.toString($key);
-    this.type = $type || 'text';
-
+    this.$type = $type || 'text';
     this.$input = $try($input, this.$input);
     this.$output = $try($output, this.$output);
 
-    this.$initial = parseInput(this.$input, {
+    this.$value = parseInput(this.$input, {
       isEmptyArray,
       type: this.type,
-      unified: checkArray($data),
-      separated: checkArray($value),
+      unified: $data,
+      separated: $props.$value,
+    });
+
+    this.$initial = parseInput(this.$input, {
+      nullable: true,
+      isEmptyArray,
+      type: this.type,
+      unified: $data,
+      separated: $props.$initial,
+      fallback: this.$value,
     });
 
     this.$default = setupDefaultProp(this, $data, $props, update, {
-      isEmptyArray, checkArray,
+      isEmptyArray,
     });
 
     setupFieldProps(this, $props, $data);
@@ -455,23 +479,35 @@ export const prototypes = {
     return toJS(val);
   },
 
-  checkDVRValidationPlugin() {
+  checkValidationPlugins() {
     const { drivers } = this.state.form.validator;
+    const form = this.state.form.name ? `${this.state.form.name}/` : '';
+
     if (_.isNil(drivers.dvr) && !_.isNil(this.rules)) {
-      // eslint-disable-next-line
-      console.warn(
-        'The DVR validation rules are defined',
-        'but no plugin provided (DVR). Field:',
-        this.path,
-      );
+      throw new Error(`The DVR validation rules are defined but no DVR plugin provided. Field: "${form + this.path}".`);
+    }
+
+    if (_.isNil(drivers.vjf) && !_.isNil(this.validators)) {
+      throw new Error(`The VJF validators functions are defined but no VJF plugin provided. Field: "${form + this.path}".`);
     }
   },
 
   @action
   initNestedFields(field, update) {
     const fields = _.isNil(field) ? null : field.fields;
-    if (_.isArray(fields)) this.hasInitialNestedFields = true;
+
+    if (_.isArray(fields) && !_.isEmpty(fields)) {
+      this.hasInitialNestedFields = true;
+    }
+
     this.initFields({ fields }, update);
+
+    if (!update && _.isArray(fields) && _.isEmpty(fields)) {
+      if (_.isArray(this.value) && !_.isEmpty(this.value)) {
+        this.hasInitialNestedFields = true;
+        this.initFields({fields, values: this.value}, update)
+      }
+    }
   },
 
   @action
@@ -512,11 +548,12 @@ export const prototypes = {
     this.$clearing = true;
     this.$touched = false;
     this.$changed = false;
+    this.$blurred = false;
 
     this.$value = defaultClearValue({ value: this.$value });
     this.files = undefined;
 
-    if (deep) this.each(field => field.clear(true, false));
+    if (deep) this.each(field => field.clear(true));
 
     this.validate({
       showErrors: this.state.options.get('showErrorsOnClear', this),
@@ -528,13 +565,14 @@ export const prototypes = {
     this.$resetting = true;
     this.$touched = false;
     this.$changed = false;
+    this.$blurred = false;
 
     const useDefaultValue = (this.$default !== this.$initial);
     if (useDefaultValue) this.value = this.$default;
     if (!useDefaultValue) this.value = this.$initial;
     this.files = undefined;
 
-    if (deep) this.each(field => field.reset(true, false));
+    if (deep) this.each(field => field.reset(true));
 
     this.validate({
       showErrors: this.state.options.get('showErrorsOnReset', this),
@@ -583,6 +621,20 @@ export const prototypes = {
           this.debouncedValidation({
             showErrors: opt.get('showErrorsOnChange', this),
           }));
+    } else if (
+      opt.get('validateOnChangeAfterInitialBlur', this) ||
+      opt.get('validateOnChangeAfterSubmit', this)
+    ) {
+      this.disposeValidationOnChange = observe(this, '$value', () => (
+        !this.actionRunning &&
+        (
+          (opt.get('validateOnChangeAfterInitialBlur', this) && this.blurred) ||
+          (opt.get('validateOnChangeAfterSubmit', this) && this.state.form.submitted)
+        ) &&
+        this.debouncedValidation({
+          showErrors: opt.get('showErrorsOnChange', this),
+        })
+      ));
     }
   },