+++ src/index.js
@@ -1,4 +1,4 @@
-import { useStrict } from 'mobx';
+import { useStrict, configure } from 'mobx';
 
 import Form, { prototypes as formPrototypes } from './Form';
 import Field, { prototypes as fieldPrototypes } from './Field';
@@ -10,15 +10,17 @@ import fieldActions from './shared/Actions';
 import fieldUtils from './shared/Utils';
 import fieldEvents from './shared/Events';
 
-const { TEST } = process.env;
-
 /**
   Enables MobX strict mode globally (TEST only).
   - - - - - - - - - - - - - - - - - -
   In strict mode, it is not allowed to
   change any state outside of an action
 */
-if (TEST) useStrict(true);
+if (process.env.TEST) {
+  configure
+    ? configure({ enforceActions: true })
+    : useStrict(true);
+}
 
 /**
   Extend Classes with Prototype