+++ src/utils.js
@@ -1,27 +1,17 @@
 import _ from 'lodash';
+import { values as mobxValues, keys as mobxKeys } from 'mobx';
 
-const props = {
-  booleans: ['hasError', 'isValid', 'isDirty', 'isPristine', 'isDefault', 'isEmpty', 'focused', 'touched', 'changed', 'disabled', 'resetting', 'clearing'],
-  field: ['value', 'initial', 'default', 'label', 'placeholder', 'disabled', 'related', 'options', 'extra', 'bindings', 'type', 'hooks', 'handlers', 'error'],
-  separated: ['values', 'initials', 'defaults', 'labels', 'placeholders', 'disabled', 'related', 'options', 'extra', 'bindings', 'types', 'hooks', 'handlers'],
-  handlers: ['onChange', 'onToggle', 'onFocus', 'onBlur', 'onDrop', 'onSubmit', 'onReset', 'onClear', 'onAdd', 'onDel'],
-  function: ['observers', 'interceptors', 'input', 'output'],
-  validation: ['rules', 'validators', 'validateWith'],
-  types: {
-    isDirty: 'some',
-    isValid: 'every',
-    isPristine: 'every',
-    isDefault: 'every',
-    isEmpty: 'every',
-    hasError: 'some',
-    focused: 'some',
-    touched: 'some',
-    changed: 'some',
-    disabled: 'every',
-    clearing: 'every',
-    resetting: 'every',
-  },
-};
+import props from './props';
+
+const getObservableMapValues = observableMap =>
+  mobxValues
+    ? mobxValues(observableMap)
+    : observableMap.values();
+
+const getObservableMapKeys = observableMap =>
+  mobxValues
+    ? mobxKeys(observableMap)
+    : observableMap.keys();
 
 const checkObserveItem = change => ({
   key, to, type, exec,
@@ -63,6 +53,7 @@ const hasProps = ($type, $data) => {
     ]; break;
     default: $props = null;
   }
+
   return _.intersection($data, $props).length > 0;
 };
 
@@ -98,10 +89,8 @@ const hasSome = (obj, keys) =>
 const isPromise = obj => (!!obj && typeof obj.then === 'function'
   && (typeof obj === 'object' || typeof obj === 'function'));
 
-const isStruct = ({ fields }) => (
-  _.isArray(fields) &&
-  _.every(fields, _.isString)
-);
+const isStruct = struct =>
+  (_.isArray(struct) && _.every(struct, _.isString));
 
 const isEmptyArray = field =>
   (_.isEmpty(field) && _.isArray(field));
@@ -130,7 +119,7 @@ const allowNested = (field, strictProps) =>
     ]) || strictProps);
 
 const parseIntKeys = fields =>
-  _.map(fields.keys(), _.ary(parseInt, 1));
+  _.map(getObservableMapKeys(fields), _.ary(_.toNumber, 1));
 
 const hasIntKeys = fields =>
   _.every(parseIntKeys(fields), _.isInteger);
@@ -188,4 +177,6 @@ export default {
   $hasFiles,
   $isBool,
   $try,
+  getObservableMapKeys,
+  getObservableMapValues,
 };