+++ tests/data/forms/fixes/form.o.js
@@ -1,6 +1,6 @@
 import { Form } from '../../../../src';
 
-const fields = [
+const struct = [
   'roles[]',
   'roles[].id',
   'array[]',
@@ -12,11 +12,21 @@ class NewForm extends Form {
   hooks() {
     return {
       onInit() {
-        this.$('array').add();
+        this.$('array').add(); // 0
+        this.$('array').add(); // 1
+        this.$('array').add(); // 2
+
         this.$('array').del(0);
+
+        const proxy = new Proxy({
+          preventDefault: () => {},
+        }, {});
+
+        this.$('array').onDel(proxy, 1);
+        this.$('array[2]').onDel(proxy);
       },
     };
   }
 }
 
-export default new NewForm({ fields }, { name: 'Fixes-O' });
+export default new NewForm({ struct }, { name: 'Fixes-O' });