+++ tests/data/forms/nested/form.u.js
@@ -1,12 +1,13 @@
 import { expect } from 'chai';
 import { Form } from '../../../../src';
 import { isEmail, shouldBeEqualTo } from '../../extension/vjf';
+import vjf from '../../../../src/validators/VJF';
 
 const fields = [
   'user',
   // TO FIX
-  // 'user.email',
-  // 'user.emailConfirm',
+  'user.email',
+  'user.emailConfirm',
 ];
 
 const labels = {
@@ -33,6 +34,12 @@ const related = {
 };
 
 class NewForm extends Form {
+
+  plugins() {
+    return {
+      vjf: vjf(),
+    };
+  }
 
   options() {
     return {